Detect core count more reliably and leverage all of them on HPC
detectCores() returns the number of CPU cores installed on a
computer, not the number of cores that are available to be used by
the running processes. On HPC clusters, the available cores are
often restricted by linux control groups to some lesser number
than the total installed on the system. availableCores() is
control group aware and will return the correct number on HPC
clusters as well as other systems.

Create variable HPC with default value FALSE for determining if
running on HPC system.

On HPC systems, set the makePSOCKcluster size to the number of
cores, reserving no additional cores. Reserving additional cores
for system stability is unnessesary within an HPC environment.